+++ nylas/handler/api_resources.py
@@ -50,6 +50,7 @@ class CreatableApiResource(Resource):
         request_body=None,
         overrides=None,
     ) -> Response:
+
         response_json, response_headers = self._http_client._execute(
             "POST", path, headers, query_params, request_body, overrides=overrides
         )